+++ Source/Application/Scenes/Objects/Sphere.hpp
@@ -20,7 +20,7 @@ struct Sphere : public Cookable
     void unbind();
 
     void draw(const Camera& camera, const glm::vec3& position = {}, const glm::vec3& orientation = {}, const std::vector<std::unique_ptr<Light>>& lights = {});
-    void drawBatch(int amount, const Camera& camera);
+    void drawBatch(size_t amount, const Camera& camera);
 
     std::shared_ptr<BaseShape> shape();
